Added system block diagrams
There are two diagrams - one for Android and the other for the Raspberry Pi. From a high level, they are identical other than terminology. Android uses Activities to separate functionality; Python on the Raspberry Pi will use Threads. The operating systems are also different, which may have an impact on our design and implementation.